requirements.txt
numpy==1.24.2
opencv-python==4.4.0.42
pandas
scikit-learn
tensorflow==2.10.0
segmentation_models==1.0.1
torch 
torchvision 
timm 
matplotlib 
Pillow 
tqdm
seaborn
